update of: input&outputDto, EveryTransfer, ServiceUpdate, ModelUpdate, Exceptions cleaned and new handler, Fixed controllers

+++ src/main/java/com/digicoachindezorg/didz_backend/models/ContactForm.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class ContactForm {
     @Id
     @GeneratedValue
-    private Integer contactFormId;
+    private Long contactFormId;
     private String companyName;
     private String name;
     private Integer phoneNumber;
