+++ src/main/java/com/digicoachindezorg/didz_backend/models/Invoice.java
@@ -18,7 +18,7 @@ import java.util.List;
 public class Invoice {
     @Id
     @GeneratedValue
-    private Integer invoiceId;
+    private Long invoiceId;
     private LocalDate orderDate;
     private Double totalPrice;
     private String address;
